chore: extract protobq.Load implementation to load.go
Makes it easier to work with the load call hierarchy when it's all
linear in the same file.
